[refactor] RacingResultTest, Test 추가
